Rename and move jsonl_input to connector.jsonl.input
