+++ logprep/connector/connector_factory.py
@@ -9,7 +9,7 @@ from logprep.connector.connector_factory_error import (
 from logprep.input.input import Input
 from logprep.output.output import Output
 from logprep.connector.dummy.input import DummyInput
-from logprep.input.jsonl_input import JsonlInput
+from logprep.connector.jsonl.input import JsonlInput
 from logprep.connector.json.input import JsonInput
 from logprep.connector.confluent_kafka.input import (
     ConfluentKafkaInput,
