+++ logprep/connector/jsonl/input.py
